+++ snowflake-distributed-service-impl/src/main/java/com/dwarfeng/sfds/impl/service/telqos/GenerateCommand.java
@@ -7,6 +7,7 @@ import com.dwarfeng.springtelqos.stack.command.Context;
 import com.dwarfeng.springtelqos.stack.exception.TelqosException;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
+import org.apache.commons.lang3.tuple.Pair;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -16,22 +17,27 @@ import java.util.List;
 import java.util.Objects;
 
 @TelqosCommand
-public class GenCommand extends CliCommand {
+public class GenerateCommand extends CliCommand {
 
     private static final String IDENTITY = "gen";
-    private static final String DESCRIPTION = "生成ID并导出到CSV操作";
+    private static final String DESCRIPTION = "生成 ID 并导出到 CSV 操作";
 
-    private static final String CMD_OPTION_SIZE = "s";
-    private static final String CMD_OPTION_FILE = "f";
+    private static final String COMMAND_OPTION_GENERATE = "generate";
 
-    private static final String CMD_LINE_SYNTAX_INTERACTIVELY = IDENTITY;
-    private static final String CMD_LINE_SYNTAX_NORMAL = IDENTITY + " "
-            + CommandUtil.concatOptionPrefix(CMD_OPTION_SIZE) + " [size] ["
-            + CommandUtil.concatOptionPrefix(CMD_OPTION_FILE) + " [file-path]]";
+    private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_GENERATE
+    };
+
+    private static final String COMMAND_OPTION_SIZE = "s";
+    private static final String COMMAND_OPTION_FILE = "f";
+
+    private static final String CMD_LINE_SYNTAX_GENERATE = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_GENERATE) + " [" +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_SIZE) + " size] [" +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_FILE) + " file-path]";
 
     private static final String[] CMD_LINE_ARRAY = new String[]{
-            CMD_LINE_SYNTAX_INTERACTIVELY,
-            CMD_LINE_SYNTAX_NORMAL
+            CMD_LINE_SYNTAX_GENERATE
     };
 
     private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
@@ -41,7 +47,7 @@ public class GenCommand extends CliCommand {
 
     private final GenerateQosService generateQosService;
 
-    public GenCommand(GenerateQosService generateQosService) {
+    public GenerateCommand(GenerateQosService generateQosService) {
         super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
         this.generateQosService = generateQosService;
     }
@@ -49,44 +55,67 @@ public class GenCommand extends CliCommand {
     @Override
     protected List<Option> buildOptions() {
         List<Option> list = new ArrayList<>();
-        list.add(Option.builder(CMD_OPTION_SIZE).optionalArg(true).hasArg(true).type(Number.class)
-                .argName("size").desc("生成数量").build());
-        list.add(Option.builder(CMD_OPTION_FILE).optionalArg(true).hasArg(true).type(String.class)
-                .argName("file-path").desc("生成的CSV的路径").build());
+        list.add(Option.builder(COMMAND_OPTION_GENERATE).desc("生成 ID").build());
+        list.add(
+                Option.builder(COMMAND_OPTION_SIZE).desc("生成数量").hasArg().type(Number.class).build()
+        );
+        list.add(
+                Option.builder(COMMAND_OPTION_FILE).desc("生成的 CSV 的路径").hasArg().type(String.class).build()
+        );
         return list;
     }
 
+    @SuppressWarnings("SwitchStatementWithTooFewBranches")
     @Override
     protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
         try {
-            int size;
-            if (cmd.hasOption(CMD_OPTION_SIZE)) {
-                Number sizeNumber = (Number) cmd.getParsedOptionValue(CMD_OPTION_SIZE);
-                if (Objects.isNull(sizeNumber)) {
-                    size = DEFAULT_SIZE;
-                } else {
-                    size = sizeNumber.intValue();
-                }
-            } else {
-                context.sendMessage("请输入需要生成的 ID 的数量:");
-                size = Integer.parseInt(context.receiveMessage());
+            Pair<String, Integer> pair = CommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
+            if (pair.getRight() != 1) {
+                context.sendMessage(CommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
+                context.sendMessage(CMD_LINE_SYNTAX);
+                return;
             }
-            List<Long> longs = generateQosService.generateLong(size);
-
-            if (cmd.hasOption(CMD_OPTION_FILE)) {
-                print2File(longs, context, cmd);
-            } else {
-                print2Screen(longs, context);
+            switch (pair.getLeft()) {
+                case COMMAND_OPTION_GENERATE:
+                    handleGenerate(context, cmd);
+                    break;
+                default:
+                    throw new IllegalStateException("不应该执行到此处, 请联系开发人员");
             }
         } catch (Exception e) {
             throw new TelqosException(e);
         }
     }
 
+    private void handleGenerate(Context context, CommandLine cmd) throws Exception {
+        int size = parseSize(context, cmd);
+        List<Long> longs = generateQosService.generateLong(size);
+
+        if (cmd.hasOption(COMMAND_OPTION_FILE)) {
+            print2File(longs, context, cmd);
+        } else {
+            print2Screen(longs, context);
+        }
+    }
+
+    private int parseSize(Context context, CommandLine cmd) throws Exception {
+        if (cmd.hasOption(COMMAND_OPTION_SIZE)) {
+            Number sizeNumber = (Number) cmd.getParsedOptionValue(COMMAND_OPTION_SIZE);
+            if (Objects.isNull(sizeNumber)) {
+                return DEFAULT_SIZE;
+            } else {
+                return sizeNumber.intValue();
+            }
+        } else {
+            context.sendMessage("请输入需要生成的 ID 的数量:");
+            return Integer.parseInt(context.receiveMessage());
+        }
+    }
+
     private void print2File(List<Long> longs, Context context, CommandLine cmd) throws Exception {
         File file;
 
-        String optionFileName = (String) cmd.getParsedOptionValue(CMD_OPTION_FILE);
+        String optionFileName = (String) cmd.getParsedOptionValue(COMMAND_OPTION_FILE);
         if (Objects.nonNull(optionFileName)) {
             file = new File((optionFileName));
             if (file.isDirectory()) {
@@ -119,7 +148,7 @@ public class GenCommand extends CliCommand {
             }
             bw.flush();
         }
-        context.sendMessage("CSV文件已导出至 " + file.getAbsolutePath());
+        context.sendMessage("CSV 文件已导出至 " + file.getAbsolutePath());
     }
 
     private void print2Screen(List<Long> longs, Context context) throws Exception {